requirements.txt
flask
google-cloud-storage
gunicorn
google-cloud-storage
google-cloud-firestore
selenium==4.34.0
webdriver-manager==4.0.2
yt-dlp==2023.6.3
